Move dt_start out of RRuleProperties. Fixes #45

+++ rrule/src/iter/yearinfo.rs
@@ -170,18 +170,14 @@ pub(crate) fn rebuild_year(
         // days from last year's last week number in
         // this year.
         let lnum_weeks = if !properties.by_week_no.iter().any(|&week_no| week_no == -1) {
-            let lyear_weekday = Utc.ymd(year - 1, 1, 1).weekday().num_days_from_monday() as u8;
+            let lyear_weekday = Utc.ymd(year - 1, 1, 1).weekday().num_days_from_monday() as isize;
 
-            let ln_no1_week_start = pymod(
-                7 - lyear_weekday as isize + properties.week_start as isize,
-                7,
-            );
+            let ln_no1_week_start = pymod(7 - lyear_weekday + properties.week_start as isize, 7);
 
-            let lyear_len = get_year_len(year - 1);
+            let lyear_len = get_year_len(year - 1) as isize;
             let week_start = if ln_no1_week_start >= 4 {
                 //ln_no1_week_start = 0;
-                lyear_len as isize
-                    + pymod(lyear_weekday as isize - properties.week_start as isize, 7)
+                lyear_len + pymod(lyear_weekday - properties.week_start as isize, 7)
             } else {
                 year_len as isize - no1_week_start
             };
